fix artifactory mem files
