+++ api/api.go
@@ -1,9 +1,7 @@
 package api
 
 import (
-	"context"
 	"encoding/json"
-	"io/fs"
 	"net/http"
 	"os"
 	"strconv"
@@ -114,13 +112,7 @@ func New(options *Options) *API {
 	r.Post("/api/extensionquery", api.extensionQuery)
 
 	// Endpoint for getting an extension's files or the extension zip.
-	r.Mount("/files", http.StripPrefix("/files", http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
-		// Convert the storage into a fs.FS, including the request context
-		http.FileServerFS(&contextFs{
-			ctx:  r.Context(),
-			open: options.Storage.Open,
-		}).ServeHTTP(rw, r)
-	})))
+	r.Mount("/files", http.StripPrefix("/files", storage.HTTPFileServer(options.Storage)))
 
 	// VS Code can use the files in the response to get file paths but it will
 	// sometimes ignore that and use requests to /assets with hardcoded types to
@@ -263,13 +255,4 @@ func (api *API) assetRedirect(rw http.ResponseWriter, r *http.Request) {
 	}
 
 	http.Redirect(rw, r, url, http.StatusMovedPermanently)
-}
-
-type contextFs struct {
-	ctx  context.Context
-	open func(ctx context.Context, name string) (fs.File, error)
-}
-
-func (c *contextFs) Open(name string) (fs.File, error) {
-	return c.open(c.ctx, name)
 }
